🚚 move verification file to algorithms/
