+++ src/dd/CMakeLists.txt
@@ -19,7 +19,6 @@ if(NOT TARGET ${MQT_CORE_TARGET_NAME}-dd)
     RealNumber.cpp
     RealNumberUniqueTable.cpp
     Simulation.cpp
-    Verification.cpp
     statistics/MemoryManagerStatistics.cpp
     statistics/Statistics.cpp
     statistics/TableStatistics.cpp
