+++ test/CMakeLists.txt
@@ -16,6 +16,7 @@ package_add_test(
   algorithms/eval_dynamic_circuits.cpp
   algorithms/test_qft.cpp
   algorithms/test_grover.cpp
+  algorithms/test_partial_equivalence.cpp
   algorithms/test_bernsteinvazirani.cpp
   algorithms/test_entanglement.cpp
   algorithms/test_grcs.cpp
